I changed the functions of accessibility when opening lists on the footer to using useState hook.
aria-hidden = {!isLargeScreen && !isListOpen}
means that aria-hidden will be 'false' when
the list is opened and it surpasses 1200px
media query. Default value of isListOpen(1-4)
is false.

I pass the state and setState when pressing
the btn on mobiles, it will change the state
to true or false, then this changes the
attribute [data-footer-list-expanded] of the
ul '.footer-section__list' to open or close.

Also the aria-expandex depends on
isListOpen(1-4) to change.